requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
pandas>=2.0,<3.0
python-dotenv>=1.0,<2.0
gspread>=5.12,<6.0
oauth2client>=4.1,<5.0
reportlab>=4.0,<5.0
openpyxl>=3.1,<4.0
plotly>=5.0,<6.0
kaleido>=0.2.1,<0.3.0
